% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{adding_age_preds_to_pdata}
\alias{adding_age_preds_to_pdata}
\title{Title of the function}
\usage{
adding_age_preds_to_pdata(pdata, mat_t, REG = TRUE, PASTA = TRUE, CT46 = FALSE)
}
\description{
Description of the function.
}
