% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_gr_from_gtf}
\alias{annotate_gr_from_gtf}
\title{Annotate Granges from GTF}
\usage{
annotate_gr_from_gtf(
  gr,
  gtf_gr = NULL,
  gtf_TxDb,
  genome = NULL,
  hard_extension = 10,
  extension = 2000,
  invert_strand = FALSE,
  annotationType = "any",
  transcriptDetails = TRUE,
  SequenceAnalysis = TRUE,
  PAS_Types = FALSE,
  pA_motif_max_position = 60,
  AAA_motif_min_position = 10,
  polystretch_length = 10,
  max_mismatch = 1
)
}
\description{
importFrom
}
