% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model_preprocessing.R
\name{applying_rank_normalization}
\alias{applying_rank_normalization}
\title{Apply Rank Normalization to Matrix}
\usage{
applying_rank_normalization(mat)
}
\arguments{
\item{mat}{Matrix. Count matrix.}
}
\value{
Matrix. Rank-normalized matrix.
}
\description{
Applies rank normalization across each column of the matrix.
}
