% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_C46}
\alias{beta_C46}
\title{Beta coefficient for conveting Pasta predictions into age-shift units}
\format{
A cv.glmnet created using the glmnet package.
}
\usage{
beta_C46
}
\description{
The Pasta model is a glmnet model which was trained on transcriptomic samples of 21 datasets and multiple tissues to predict the age-shift between two pairs of samples.
}
\keyword{datasets}
