% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcAvgTandemUTRlength}
\alias{CalcAvgTandemUTRlength}
\title{Calculate average tandem UTR length for each cell}
\usage{
CalcAvgTandemUTRlength(
  object,
  assay = "pA",
  genes = NULL,
  pseudocount.use = 1,
  min.reads.per.cell = 1
)
}
\arguments{
\item{assay}{Name of polyAsite.Assay}

\item{genes}{Vector of Gene names containing tandem UTRs. If Null, all annotated tandem UTRs are used (default).}

\item{min.reads.per.cell}{Use genes with minimum average read count (default 1).}

\item{pseudocount}{Pseudocount to add for tandem UTR length calculation (default 0).}
}
\value{
Calculates the average tandem UTR length for each cell
}
\description{
Calculate average tandem UTR length for each cell
}
\concept{utilities}
