% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalculateTandemUTRlength}
\alias{CalculateTandemUTRlength}
\title{Calculate Tandem UTR Length Wrapper}
\usage{
CalculateTandemUTRlength(
  object = NULL,
  assay = NULL,
  genes = NULL,
  pseudocount = 0,
  min.reads.per.cell = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Name of polyAsite.Assay}

\item{genes}{Vector of Gene names containing tandem UTRs. If Null, all annotated tandem UTRs are used (default).}

\item{pseudocount}{Pseudocount to add for tandem UTR length calculation (default 0).}

\item{min.reads.per.cell}{Use genes with minimum average read count (default 1).}
}
\value{
Returns a matrix with average tandem UTR length
}
\description{
Calculate avergae length of tandem 3'UTRs
}
\concept{utilities}
