% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcUTRlength}
\alias{calcUTRlength}
\title{Calculate Tandem UTR Length per gene}
\usage{
calcUTRlength(gene, df, object, assay = assay, pc = 0, Min.Reads.Per.Cell = 1)
}
\arguments{
\item{gene}{Name of gene with tandem UTR.}

\item{object}{Seurat object}

\item{assay}{Name of polyAsite.Assay}

\item{pc}{Pseudocount to add for tandem UTR length calculation (default 0).}

\item{Min.Reads.Per.Cell}{Use genes with minimum average read count (default 1).}

\item{N.cells}{Number of cells}
}
\value{
Returns a vector with average tandem UTR length
}
\description{
Calculate avergae length of tandem 3'UTRs
}
\concept{utilities}
