% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_conversion.R
\name{check_biomart_attributes}
\alias{check_biomart_attributes}
\title{Check biomaRt Attributes}
\usage{
check_biomart_attributes(cur_mart, pattern)
}
\arguments{
\item{cur_mart}{Mart object.}

\item{pattern}{Character. Regular expression to filter attributes.}
}
\value{
Printed data frame of matching attributes.
}
\description{
Lists attributes from a biomaRt Mart filtered by a pattern.
}
