% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{checking_if_url_is_valid}
\alias{checking_if_url_is_valid}
\title{Title of the function}
\usage{
checking_if_url_is_valid(url_in, t = 2)
}
\description{
Description of the function.
}
\keyword{internal}
