% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef_agediff}
\alias{coef_agediff}
\title{Coefficients for the PASTA model}
\format{
A cv.glmnet created using the glmnet package.
}
\usage{
coef_agediff
}
\description{
The PASTA model is a glmnet model which was trained on transcriptomic samples of 21 datasets and multiple tissues to predict the age-shift between two pairs of samples.
}
\keyword{datasets}
