% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_age_shift.R
\name{collapse_matrix_by_vector}
\alias{collapse_matrix_by_vector}
\title{A function to collapse replicates for the same control samples.}
\usage{
collapse_matrix_by_vector(mat_ctl0, vec)
}
\arguments{
\item{mat_ctl0}{Count matrix for the control samples.}

\item{vec}{Character vector which indicates the label of each columns. Controls with the same label will be aggregated.}
}
\value{
Returns the input and print its length.
}
\description{
A function to collapse replicates for the same control samples.
}
\examples{
library(magrittr)
mat_ctl <- matrix(sample(c(rep(0, 10), 1:14)), nrow = 4, ncol = 6)
colnames(mat_ctl) = paste0('ctl_', 1:6)
rownames(mat_ctl) = paste0('gene_', 1:4)
set.seed(1)
mat_ctl0 = rank_normalize_mat(mat_ctl) \%>\% t
vec = rep(c('ctl_A', 'ctl_B', 'ctl_C'), 2)
collapse_matrix_by_vector(mat_ctl0, vec)
}
