% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_conversion.R
\name{converting_entrez_to_ensembl_gene_ids}
\alias{converting_entrez_to_ensembl_gene_ids}
\title{Convert Entrez to Ensembl Gene IDs (using GEO names)}
\usage{
converting_entrez_to_ensembl_gene_ids(mat)
}
\arguments{
\item{mat}{Matrix. Count matrix with original gene identifiers.}
}
\value{
Matrix. Count matrix with Ensembl gene IDs as row names.
}
\description{
Replaces row names in the matrix with Ensembl gene IDs from a GEO-specific mapping.
}
