% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{converting_gene_ids_to_ensembl_gene_ids}
\alias{converting_gene_ids_to_ensembl_gene_ids}
\title{Title of the function}
\usage{
converting_gene_ids_to_ensembl_gene_ids(
  mat,
  cur_mart,
  platform = "illumina_humanht_12_v4"
)
}
\description{
Description of the function.
}
