% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{converting_gene_ids_to_ensembl_gene_ids_with_biomaRt}
\alias{converting_gene_ids_to_ensembl_gene_ids_with_biomaRt}
\title{Title of the function}
\usage{
converting_gene_ids_to_ensembl_gene_ids_with_biomaRt(mat)
}
\description{
Description of the function.
}
