% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvfit_C46}
\alias{cvfit_C46}
\title{The Traditional classifier <40 vs >60 years old model}
\format{
A cv.glmnet created using the glmnet package.
}
\usage{
cvfit_C46
}
\description{
The Traditional classifier model is a glmnet model which was trained on transcriptomic samples of 21 datasets and multiple tissues to predict if a sample is young or old.
}
\keyword{datasets}
