% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvfit_pasta}
\alias{cvfit_pasta}
\title{Te PASTA model}
\format{
A cv.glmnet created using the glmnet package.
}
\usage{
cvfit_pasta
}
\description{
The PASTA model is a glmnet model which was trained on transcriptomic samples of more than 10 datasets and multiple tissues to predict the age-shift between two pairs of samples. It is highly predictive of whether one sample is younger than another sample.
}
\keyword{datasets}
