% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ES_GSE103938}
\alias{ES_GSE103938}
\title{Test dataset GSE103938}
\format{
A data frame with 57232 genes and 21 samples
}
\source{
The object was obtained from the supplementary file GSE103938_GeneCounts.xlsx at the link: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE103938.
}
\usage{
ES_GSE103938
}
\description{
The raw counts ExpressionSet R object for GSE103938 obtained from GEO. Title of the dataset on GEO: "Transcriptome analysis of OSKM- or RAS-induced senescent IMR90 fibroblasts treated with Rapamycin.".
}
\keyword{datasets}
