% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FeaturesToGRanges}
\alias{FeaturesToGRanges}
\title{Features to GRanges}
\usage{
FeaturesToGRanges(regions, sep = c(":", ",", ":"))
}
\arguments{
\item{regions}{Vector of genomic coordinate ranges in format chr:start,end:strand}

\item{sep}{Vector of separators to use for genomic ranges string. First element separates chromosome
and start position, second element separates start position and end end position,
third element separates end position and strand.}
}
\value{
Returns a genomic ranges
}
\description{
Transform List of Features to GRanges object
}
\concept{utilities}
