% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{filter_cell_types_in_seu_object}
\alias{filter_cell_types_in_seu_object}
\title{Title of the function}
\usage{
filter_cell_types_in_seu_object(
  seu,
  n_cell_min = 500,
  dry_run = FALSE,
  verbose = FALSE
)
}
\description{
Description of the function.
}
