% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model_preprocessing.R
\name{filtering_age_model_genes}
\alias{filtering_age_model_genes}
\title{Filter Age Model Genes from Count Matrix}
\usage{
filtering_age_model_genes(mat)
}
\arguments{
\item{mat}{Matrix. Count matrix.}
}
\value{
Matrix. Filtered count matrix with median imputation.
}
\description{
Subsets the count matrix to include only genes used in the age prediction model.
}
