% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{filtering_age_model_genes_and_rank_norm}
\alias{filtering_age_model_genes_and_rank_norm}
\title{Title of the function}
\usage{
filtering_age_model_genes_and_rank_norm(mat)
}
\description{
Description of the function.
}
