% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{FindVariableFeatures.polyAsiteAssay}
\alias{FindVariableFeatures.polyAsiteAssay}
\title{Finds variable features based on polyA residuals}
\usage{
\method{FindVariableFeatures}{polyAsiteAssay}(
  object,
  nfeatures = 2000,
  selection.method = "residuals",
  gene.names = "symbol",
  min.variance = 0.5,
  ...
)
}
\arguments{
\item{object}{Object containing a polyAsite assay}

\item{nfeatures}{Number of features to select as top variable features.}

\item{selection.method}{How to calculate polyA residuals. If "residuals" (default),
will rank all polyA sites by their variance and pick at most 1 polyA site per gene.
Otherwise, will use Seurat FindVariableFeatures functions.}

\item{gene.names}{Column in meta features that contains gene names. At most one feature per gene will be selected.}

\item{min.variance}{Minimum variance threshold for residuals. Only features with variance
above this threshold will be considered variable. Default is 0.5.}

\item{...}{Arguments passed to other methods}
}
\description{
Finds variable features based on polyA residuals
}
\seealso{
\code{\link[Seurat]{FindVariableFeatures}} for additional parameters.
}
\concept{preprocessing}
