% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{get_cor_by_chunk_from_pdata_big}
\alias{get_cor_by_chunk_from_pdata_big}
\title{Title of the function}
\usage{
get_cor_by_chunk_from_pdata_big(pdata_big)
}
\description{
Description of the function.
}
\examples{
library(magrittr)
library(jsutil)
data(seu_orozco_2020_retina_horizontal_cells)
seu = seu_orozco_2020_retina_horizontal_cells \%T>\% pdim # 57596 1875
rm(seu_orozco_2020_retina_horizontal_cells)
seu$age  = seu$development_stage \%>\% gsub('-year.*', '', .) \%>\% gsub('-', ' ', .)
seu$type = seu$cell_type \%>\% as.character
seu \%>\% filter_cell_types_in_seu_object(dry_run = TRUE, verbose = TRUE)
seu \%<>\% filter_cell_types_in_seu_object \%T>\% pdim # 57596 1875
seu_bulk = making_pseudobulks_from_seurat(seu)
( pdata = predicting_age_from_pseudobulks(seu_bulk) )
( pdata = making_pseudobulks_and_predict_age(seu) )
( pdata_big = predicting_age_multiple_chunks(seu) )
get_cor_by_chunk_from_pdata_big(pdata_big)
}
