% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_conversion.R
\name{get_mart_ensembl_human}
\alias{get_mart_ensembl_human}
\title{Get Ensembl Human Mart}
\usage{
get_mart_ensembl_human()
}
\value{
Mart object.
}
\description{
Returns a biomaRt Mart object for human gene annotation.
}
