% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{GetBackgroundDist}
\alias{GetBackgroundDist}
\title{Get Background Distribution}
\usage{
GetBackgroundDist(
  object,
  features,
  background,
  gene.names,
  assay,
  min.counts.background
)
}
\arguments{
\item{object}{Seurat object containing a polyAsiteAssay}

\item{features}{Features to include in calculation of polyA residuals.
If NULL, use all features.}

\item{background}{Identity of cells to use as background
If NULL, uses all cells combined as a background}

\item{gene.names}{Name of column containing gene annotations}

\item{assay}{Name of polyAsiteAssay to be used in calculating polyAresiduals}

\item{min.counts.background}{Features with at least this many counts in the background cells are included in calculation}
}
\value{
Returns a data frame containing all peaks within genes that have multiple polyA sites that meet min.counts.background criteria
}
\description{
Calculated Pseudobulk Ratios of Each Isoform within a gene for background distribution.
}
\concept{residuals}
