% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model_preprocessing.R
\name{getting_geo_count_mat_for_age_prediction}
\alias{getting_geo_count_mat_for_age_prediction}
\title{Get Processed GEO Count Matrix for Age Prediction}
\usage{
getting_geo_count_mat_for_age_prediction(gse_id)
}
\arguments{
\item{gse_id}{Character. GEO series identifier.}
}
\value{
Matrix. Processed count matrix.
}
\description{
Downloads and processes the GEO count matrix for age prediction.
}
