% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_ES_and_pdata_processing.R
\name{getting_GEO_ES}
\alias{getting_GEO_ES}
\title{Create ExpressionSet from GEO Data}
\usage{
getting_GEO_ES(mat, pdata)
}
\arguments{
\item{mat}{Matrix. Count matrix.}

\item{pdata}{data.frame. Phenotype data.}
}
\value{
ExpressionSet. ExpressionSet object.
}
\description{
Constructs an ExpressionSet object from a count matrix and phenotype data.
}
