% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_ES_and_pdata_processing.R
\name{getting_GEO_ES_for_age_model}
\alias{getting_GEO_ES_for_age_model}
\title{Get GEO ExpressionSet for Age Model}
\usage{
getting_GEO_ES_for_age_model(gse_id)
}
\arguments{
\item{gse_id}{Character. GEO series identifier.}
}
\value{
ExpressionSet.
}
\description{
Downloads the GEO count matrix and phenotype data and returns an ExpressionSet.
}
