% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_ES_and_pdata_processing.R
\name{getting_GEO_pdata}
\alias{getting_GEO_pdata}
\title{Retrieve GEO Phenotype Data}
\usage{
getting_GEO_pdata(gse_id, mat)
}
\arguments{
\item{gse_id}{Character. GEO series identifier.}

\item{mat}{Matrix. Count matrix (not used here but maintained for consistency).}
}
\value{
data.frame. Phenotype data.
}
\description{
Retrieves and combines phenotype data from GEO.
}
