% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_download.R
\name{getting_geo_url}
\alias{getting_geo_url}
\title{Construct GEO Download URL}
\usage{
getting_geo_url(gse_id)
}
\arguments{
\item{gse_id}{Character. GEO series identifier.}
}
\value{
Character. The URL to download the count matrix.
}
\description{
Constructs the download URL for GEO count matrix files.
}
\keyword{internal}
