% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{getting_geo_valid_url}
\alias{getting_geo_valid_url}
\title{Title of the function}
\usage{
getting_geo_valid_url(gse_id, return_boolean = F)
}
\description{
Description of the function.
}
\keyword{internal}
