% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GRangesToString}
\alias{GRangesToString}
\title{GRanges to String}
\usage{
GRangesToString(grange, sep = c(":", ",", ":"))
}
\arguments{
\item{grange}{A GRanges object containing genomic coordinates and strand}

\item{sep}{Vector of separators to use for genomic ranges string. First element separates chromosome
and start position, second element separates start position and end end position,
third element separates end position and strand}
}
\value{
Returns a character vector
}
\description{
Converts GRanges object to vector of strings.
}
\concept{utilities}
