% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%_\%}
\alias{\%_\%}
\title{Paste underscore Function}
\usage{
a \%_\% b
}
\arguments{
\item{a}{a character string}

\item{y}{a character string}
}
\description{
Pastes two strings by "_".
}
\examples{
'spaghetti' \%_\% 'sauce'
'noodle' \%_\% 'soup'
}

