% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.esaote.file_faster.R
\name{import.esaote.file_faster}
\alias{import.esaote.file_faster}
\title{Import a csv file from Esaote echograph}
\usage{
import.esaote.file_faster(file, condition = NA)
}
\arguments{
\item{file}{A csv file}

\item{condition}{The identification tag for the experiment group}
}
\value{
A object of class 'xtrain.one'
}
\description{
Import a csv file from Esaote echograph
}
\examples{
import.esaote.file_faster('ENDO_LONG_DISPL_SUBJECT1.csv')
}
