% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{making_pseudobulks_from_seurat}
\alias{making_pseudobulks_from_seurat}
\title{Title of the function}
\usage{
making_pseudobulks_from_seurat(seu, chunk_size = 1000, verbose = TRUE)
}
\description{
Description of the function.
}
