% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.polyAsiteAssay}
\alias{merge.polyAsiteAssay}
\title{Merge polyA site assays}
\usage{
\method{merge}{polyAsiteAssay}(x = NULL, y = NULL, add.cell.ids = NULL, cells = NULL, ...)
}
\description{
Merge polyA site assays
}
\concept{objects}
