% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.import.xtrain.R
\name{multi.import.xtrain}
\alias{multi.import.xtrain}
\title{Import multiple csv files from Esaote echograph}
\usage{
multi.import.xtrain(path = ".", condition = NA, apex.n = 8,
  echotype = "esaote", multicore = FALSE, cores = "ALL")
}
\arguments{
\item{path}{A directory where all the files are (character)}

\item{condition}{The identification tag for the experiment group (character)}

\item{echotype}{The type of echograph. Currently, only the esaote output are analyzed (character)}

\item{multicore}{If there is parallel importation or not (Logical)}

\item{cores}{The number of cores for parallel importation option. By default, 'ALL' used every cores.}

\item{apex}{The number of points by default for the apex (numeric)}
}
\value{
An object of class 'xtrain.multi'
}
\description{
Import multiple csv files from Esaote echograph
}
\examples{
import.esaote.file_faster('ENDO_LONG_DISPL_SUBJECT1.csv')
}
