% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%_\%}
\alias{\%&&\%}
\alias{\%&\%}
\alias{\%-\%}
\alias{\%.\%}
\alias{\%//\%}
\alias{\%_\%}
\alias{\%|\%}
\alias{pasta}
\title{Paste underscore Function}
\usage{
a \%_\% b

a \%-\% b

a \%.\% b

a \%//\% b

a \%&\% b

a \%&&\% b

a \%|\% b
}
\arguments{
\item{a, b}{characters.}
}
\description{
\%_\% pastes two strings by "_".

\%-\% pastes two strings by "-".

\%.\% pastes two strings by ".".

\%//\% pastes two strings by "/".

\%&\% pastes two strings by "".

\%&&\% pastes two strings by " ".

\%|\% pastes two strings with a "|".

Make pasta with strings
}
\examples{
'spaghetti' \%_\% 'sauce' # 'spaghetti_sauce'
'ravioli' \%-\% 'recipe' # 'ravioli-recipe'
'spaghetti' \%.\% 'code' # 'spaghetti.code'
'fussili' \%//\% 'farfalle' # 'fussili/farfalle'
'pasta' \%&\% 'sciutta' # 'pastasciutta'
'noodle' \%&&\% 'soup' # 'noodle soup'
'carbonara' \%|\% 'ragù' # 'carbonara|ragù'
}

