% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{polyAsiteAssay-class}
\alias{polyAsiteAssay-class}
\alias{polyAsiteAssay}
\title{The polyAsiteAssay class}
\description{
The polyAsiteAssay object is an extended \code{\link[Signac]{ChromatinAssay}}
and \code{\link[Seurat]{Assay}} for the storage and analysis of single-cell
polyAsite data.
}
\section{Slots}{

\describe{
\item{\code{ranges}}{A \code{\link[GenomicRanges]{GRanges}} object describing the
genomic location of features in the object}

\item{\code{motifs}}{A \code{\link{Motif}} object}

\item{\code{fragments}}{A list of \code{\link{Fragment}} objects.}

\item{\code{seqinfo}}{A \code{\link[GenomeInfoDb]{Seqinfo}} object containing basic
information about the genome sequence used.}

\item{\code{annotation}}{A  \code{\link[GenomicRanges]{GRanges}} object containing
genomic annotations}
}}

\concept{assay}
