% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_age_shift.R
\name{predict_age_shift}
\alias{predict_age_shift}
\title{A function to return an object and printing its length.}
\usage{
predict_age_shift(mat1, pdata1, ctl_column, v_ctl_values, grp_column)
}
\arguments{
\item{mat1}{Count matrix, with rownames being Ensembl gene ids.}

\item{pdata1}{Sample annotation table (i.e., phenotypic data).}

\item{ctl_column}{The name of the column to use to select the control samples.}

\item{v_ctl_values}{Samples with these values in the \code{ctl_column} will be used as controls.}

\item{grp_column}{The name of the column to use to select groups of treatment and control samples to analyze together.}
}
\value{
Returns the input and print its length.
}
\description{
A function to return an object and printing its length.
}
\examples{
library(magrittr)
library(jsutil)
library(ggplot2)
library(data.table)
data(ES_GSE103938, envir = environment())
mat1 = Biobase::exprs(ES_GSE103938)
pdata1 = data.table(sample = colnames(mat1))
pdata1[, type := gsub('_.*', '', sample)]
pdata1[, rep  := gsub('.*_', '', sample)]
ctl_column   = 'type'
v_ctl_values = 'Prolif'
grp_column   = 'rep'
pdata = predict_age_shift(mat1, pdata1, ctl_column, v_ctl_values, grp_column)
pdata[, type := factor(type, levels = 
    c('Prolif', 'RAS', 'RAS1nM', 'RAS10nM', 'OSKM', 'OSKM1nM', 'OSKM10nM') \%>\% rev)]
p1 = ggplot(pdata, aes(x = type, y = age_shift, fill = type)) + 
    geom_bar(stat = 'identity') +
    facet_grid(~rep) +
    theme_classic() + 
    coord_flip()
# pdf('test.pdf'); print(p1); dev.off()
# => Please observe that this dataset is not amazing as a control since the 
#    OSKM label is misleading as the authors annotate these samples as OSKM-
#    induced senescence. Usually, the age-shift effect of OSKM is much stronger.
}
