% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{predicting_age_from_pseudobulks}
\alias{predicting_age_from_pseudobulks}
\title{Title of the function}
\usage{
predicting_age_from_pseudobulks(
  seu_bulk,
  REG = TRUE,
  PASTA = TRUE,
  CT46 = TRUE
)
}
\description{
Description of the function.
}
