% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model_preprocessing.R
\name{processing_geo_count_mat_for_age_prediction}
\alias{processing_geo_count_mat_for_age_prediction}
\title{Process GEO Count Matrix for Age Prediction}
\usage{
processing_geo_count_mat_for_age_prediction(mat)
}
\arguments{
\item{mat}{Matrix. Raw count matrix.}
}
\value{
Matrix. Processed count matrix ready for age prediction.
}
\description{
Converts gene IDs and filters & normalizes genes for age prediction.
}
