% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_age_shift.R
\name{rank_normalize_mat}
\alias{rank_normalize_mat}
\title{A function to rank normalize the count matrix.}
\usage{
rank_normalize_mat(mat1)
}
\arguments{
\item{mat1}{Count matrix.}
}
\value{
Returns the input and print its length.
}
\description{
A function to rank normalize the count matrix.
}
\examples{
mat1 <- matrix(sample(c(rep(0, 10), 1:14)), nrow = 6, ncol = 4)
set.seed(1)
rank_normalize_mat(mat1)
}
