% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{RegDMVar}
\alias{RegDMVar}
\title{Run Dirichlet Multionmial Distribution}
\usage{
RegDMVar(
  ec,
  var,
  m,
  m.background,
  background.dist,
  gene.sum,
  background.cells,
  min.variance = min.variance
)
}
\value{
Returns a data frame containing all peaks within genes that have multiple polyA sites that meet min.counts.background criteria
}
\description{
Calculated Pseudobulk Ratios of Each Isoform within a gene for background distribution.
}
\concept{residuals}
