% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_conversion.R
\name{renaming_rows_in_mat_after_gene_id_conversion}
\alias{renaming_rows_in_mat_after_gene_id_conversion}
\title{Rename Matrix Rows after Gene ID Conversion}
\usage{
renaming_rows_in_mat_after_gene_id_conversion(mat, dt_ids)
}
\arguments{
\item{mat}{Matrix. Count matrix.}

\item{dt_ids}{data.table. Conversion table with original and Ensembl gene IDs.}
}
\value{
Matrix. Count matrix with new row names.
}
\description{
Renames the rows of a matrix based on a conversion table.
}
