% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RunPCA.polyAsiteAssay}
\alias{RunPCA.polyAsiteAssay}
\title{RunPCA only PolyA site assay}
\usage{
\method{RunPCA}{polyAsiteAssay}(
  object,
  features = NULL,
  do.scale = TRUE,
  do.center = TRUE,
  residuals.max = 10,
  residuals.min = -10,
  verbose = TRUE
)
}
\arguments{
\item{object}{A polyAsiteAssay}

\item{features}{Features to run PCA on}

\item{do.scale}{Scale residuals before performing PCA. Default is TRUE.}

\item{do.center}{Center residuals before performing PCA. Default is TRUE.}

\item{residuals.max}{Clip residuals above this value before performing PCA (default is 10)}

\item{residuals.min}{Clip residuals below this value before performing PCA (default is -10)}

\item{verbose}{Verbose output}
}
\description{
RunPCA only PolyA site assay
}
\concept{dimensional_reduction}
