% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seu_gabitto_2024_L5_ET_MTG_neuron}
\alias{seu_gabitto_2024_L5_ET_MTG_neuron}
\title{Test dataset seu_gabitto_2024_L5_ET_MTG_neuron}
\format{
A data frame with 36,412 genes and 2,590 cells.
}
\source{
The object was obtained from the subset "L5 ET - MTG: Seattle Alzheimer's Disease Atlas (SEA-AD)" at the link: https://cellxgene.cziscience.com/collections/1ca90a2d-2943-483d-b678-b809bf464c30.
}
\usage{
seu_gabitto_2024_L5_ET_MTG_neuron
}
\description{
The Seurat R object for seu_gabitto_2024_L5_ET_MTG_neuron obtained from CZ cellXGene. Title of the dataset on cellXGene.: "SEA-AD: Seattle Alzheimer’s Disease Brain Cell Atlas".
}
\keyword{datasets}
