% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seu_orozco_2020_retina_horizontal_cells}
\alias{seu_orozco_2020_retina_horizontal_cells}
\title{Test dataset seu_orozco_2020_retina_horizontal_cells}
\format{
A data frame with 57596 genes and 1875 cells.
}
\source{
The object was obtained from the supplementary file GSE103938_GeneCounts.xlsx at the link: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE103938.
}
\usage{
seu_orozco_2020_retina_horizontal_cells
}
\description{
The Seurat R object for seu_orozco_2020_retina_horizontal_cells obtained from CZ cellXGene. Title of the dataset on cellXGene.: "Integration of eQTL and a Single-Cell Atlas in the Human Eye Identifies Causal Genes for Age-Related Macular Degeneration".
}
\keyword{datasets}
