% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{subset.polyAsiteAssay}
\alias{subset.polyAsiteAssay}
\title{Subset a polyA site assay}
\usage{
\method{subset}{polyAsiteAssay}(x, features = NULL, cells = NULL, ...)
}
\arguments{
\item{x}{A polyAsiteAssay}

\item{features}{Which features to retain}

\item{cells}{Which cells to retain}
}
\description{
Subset a polyA site assay
}
\concept{objects}
