% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_genes_model}
\alias{v_genes_model}
\title{Genes used by the age-prediction models.}
\format{
Vector of ensembl gene names.
}
\usage{
v_genes_model
}
\description{
Genes used by the age-prediction models.
}
\keyword{datasets}
