% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_human_mouse_one2one}
\alias{v_human_mouse_one2one}
\title{One-to-one ortholog mapping between human and mouse genes}
\format{
A data frame or vector containing Ensemble gene identifiers for orthologous pairs
}
\usage{
v_human_mouse_one2one
}
\description{
A vector of one-to-one ortholog relationships between human and mouse genes.
}
\keyword{datasets}
