% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_new_names_geo_mat}
\alias{v_new_names_geo_mat}
\title{Genes used to convert genes ids from GEO tables from entrez to ensembl.}
\format{
Vector of ensembl gene names.
}
\usage{
v_new_names_geo_mat
}
\description{
Genes used to convert genes ids from GEO tables from entrez to ensembl.
}
\keyword{datasets}
