% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_prediction.R
\name{adding_age_preds_to_pdata}
\alias{adding_age_preds_to_pdata}
\title{Add Age Predictions to Phenotype Data}
\usage{
adding_age_preds_to_pdata(pdata, mat_t, REG = TRUE, Pasta = TRUE, CT46 = FALSE)
}
\arguments{
\item{pdata}{data.table. Phenotype data.}

\item{mat_t}{Transposed count matrix.}

\item{REG}{Logical. Include REG model predictions.}

\item{Pasta}{Logical. Include Pasta model predictions.}

\item{CT46}{Logical. Include CT46 model predictions.}
}
\value{
data.table. Updated phenotype data with age predictions.
}
\description{
Adds age prediction scores to phenotype data using different models.
}
