% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{AnnotatePASfromGTF}
\alias{AnnotatePASfromGTF}
\title{Annotate PAS from GTF file}
\usage{
AnnotatePASfromGTF(
  object,
  assay,
  gtf.file,
  genome = genome,
  hard_extension = 10,
  extension = 2000,
  invert_strand = FALSE,
  annotationType = "any",
  transcriptDetails = TRUE,
  SequenceAnalysis = TRUE,
  PAS_Types = TRUE,
  pA_motif_max_position = 60,
  AAA_motif_min_position = 10,
  polystretch_length = 13,
  max_mismatch = 1
)
}
\arguments{
\item{object}{}
}
\description{
Annotate PAS from GTF file
}
\concept{annotation}
