% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{CalcPolyAResiduals}
\alias{CalcPolyAResiduals}
\title{Calculate PolyA Residuals}
\usage{
CalcPolyAResiduals(
  object,
  assay = "polyA",
  features = NULL,
  background = NULL,
  gene.names = "symbol",
  min.counts.background = 5,
  min.variance = 0.1,
  do.center = TRUE,
  residuals.max = 10,
  residuals.min = -10,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object containing a polyAsiteAssay}

\item{assay}{Name of polyAsiteAssay to be used in calculating polyAresiduals}

\item{features}{Features to include in calculation of polyA residuals.
Default is to use all features.}

\item{background}{Identity of cells to use as background.
Default is to use all cells as a background.}

\item{gene.names}{Column containing the gene where each polyA site is annotated.
Default is symbol.}

\item{min.counts.background}{Features with at least this many counts in the background cells are included in calculation}

\item{min.variance}{Sets minimum variance. Default is 0.1.}

\item{do.center}{Return the centered residuals. Default is TRUE.}

\item{residuals.max}{Clip residuals above this value. Default is 10. If NULL, then no clipping.}

\item{residuals.min}{Clip residuals below this value. Default is -10. If NULL, then no clipping.}

\item{verbose}{Print messages.}
}
\value{
Returns a Seurat object with polyAresiduals assay
}
\description{
Calculate PolyA Residuals
}
\concept{residuals}
