% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_download.R
\name{checking_if_url_is_valid}
\alias{checking_if_url_is_valid}
\title{Check if URL is valid}
\usage{
checking_if_url_is_valid(url_in, t = 2)
}
\arguments{
\item{url_in}{Character. The URL to be checked.}

\item{t}{Numeric. Timeout in seconds.}
}
\value{
Logical. TRUE if the URL is valid, FALSE otherwise.
}
\description{
Attempts to open a connection to a URL to check its validity.
}
\keyword{internal}
