% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{converting_entrez_to_ensembl_gene_ids}
\alias{converting_entrez_to_ensembl_gene_ids}
\title{Title of the function}
\usage{
converting_entrez_to_ensembl_gene_ids(mat)
}
\description{
Description of the function.
}
