% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_conversion.R
\name{converting_gene_ids_to_ensembl_gene_ids_with_biomaRt}
\alias{converting_gene_ids_to_ensembl_gene_ids_with_biomaRt}
\title{Convert Gene IDs to Ensembl Gene IDs using biomaRt (Placeholder)}
\usage{
converting_gene_ids_to_ensembl_gene_ids_with_biomaRt(mat)
}
\arguments{
\item{mat}{Matrix. Count matrix.}
}
\value{
Matrix. Count matrix with Ensembl gene IDs.
}
\description{
This function is a placeholder for gene conversion using biomaRt.
}
