% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvfit_PASTA}
\alias{cvfit_PASTA}
\title{The PASTA model}
\format{
A cv.glmnet created using the glmnet package.
}
\usage{
cvfit_PASTA
}
\description{
The PASTA model is a glmnet model which was trained on transcriptomic samples of 21 datasets and multiple tissues to predict the age-shift between two pairs of samples.
}
\keyword{datasets}
