% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvfit_Pasta}
\alias{cvfit_Pasta}
\title{The Pasta model}
\format{
A cv.glmnet created using the glmnet package.
}
\usage{
cvfit_Pasta
}
\description{
The Pasta model is a glmnet model which was trained on transcriptomic samples of 21 datasets and multiple tissues to predict the age-shift between two pairs of samples.
}
\keyword{datasets}
