% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{DirichletMultionmial}
\alias{DirichletMultionmial}
\title{Run Dirichlet Multinomial Distribution}
\usage{
DirichletMultionmial(
  gene.test,
  background.dist,
  m.background,
  gene.sum,
  ncells
)
}
\arguments{
\item{gene.test}{which gene to use}

\item{background.dist}{dataframe containing the isoform ratios for each}

\item{m.background}{matrix of background distribution}

\item{gene.sum}{sum of count within each gene for each cell}

\item{ncells}{number of cells}
}
\value{
Returns a list where first element is matrix of expected values for each peak within the genes,
second value is matrix of variance for each peak within the gene
}
\description{
Fit dirichlet multinomial distribution on each peak within a gene using background cells.
Then calculate expected value and variance for each cell based on estimates from dirichlet multionimial regression.
}
\concept{residuals}
