% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurats_and_pseudobulks.R
\name{filter_cell_types_in_seu_object}
\alias{filter_cell_types_in_seu_object}
\title{Filter Cell Types in a Seurat Object}
\usage{
filter_cell_types_in_seu_object(
  seu,
  n_cell_min = 500,
  dry_run = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{seu}{Seurat object.}

\item{n_cell_min}{Numeric. Minimum number of cells required per cell type.}

\item{dry_run}{Logical. If TRUE, do not return filtered object.}

\item{verbose}{Logical. If TRUE, prints cell type summaries.}
}
\value{
Seurat object (if dry_run is FALSE).
}
\description{
Filters a Seurat object to include only cell types with a minimum number of cells.
}
