% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model_preprocessing.R
\name{filtering_age_model_genes_and_rank_norm}
\alias{filtering_age_model_genes_and_rank_norm}
\title{Filter and Rank-Normalize Age Model Genes}
\usage{
filtering_age_model_genes_and_rank_norm(mat)
}
\arguments{
\item{mat}{Matrix. Count matrix.}
}
\value{
Matrix. Processed count matrix.
}
\description{
Combines filtering for age model genes and rank normalization.
}
