% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_polyA.R
\name{FindDifferentialPolyA}
\alias{FindDifferentialPolyA}
\title{Find polyA sites that are differentially used across cells, utilizing
polyA residuals.}
\usage{
FindDifferentialPolyA(
  object,
  assay = "polyA",
  ident.1 = NULL,
  ident.2 = NULL,
  features = NULL,
  covariates = NULL,
  gene.names = "Gene_Symbol"
)
}
\arguments{
\item{object}{An object}

\item{assay}{name of polyAsite assay to test. Default is polyA.}

\item{ident.1}{Identity class to find polyA sites for}

\item{ident.2}{A second identity class for comparison.}

\item{features}{Features to test. Default is all features.}

\item{gene.names}{Column name providing gene annotation of each polyA site.
Default is "Gene_Symbol"}
}
\description{
Find polyA sites that are differentially used across cells, utilizing
polyA residuals.
}
\concept{differential_polyA}
