% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{FindVariableFeatures.polyAsiteAssay}
\alias{FindVariableFeatures.polyAsiteAssay}
\title{Finds variable features based on polyA residuals}
\usage{
\method{FindVariableFeatures}{polyAsiteAssay}(object, nfeatures = 2000, ...)
}
\arguments{
\item{object}{Object containing a polyAsite assay}

\item{nfeatures}{Number of features to select as top variable features.}
}
\description{
Finds variable features based on polyA residuals
}
\concept{preprocessing}
