% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{GetPolyADbAnnotation}
\alias{GetPolyADbAnnotation}
\title{Overlap PAS with polyA db}
\usage{
GetPolyADbAnnotation(
  object,
  assay = "polyA",
  polyAdb.file = NULL,
  max.dist = 50
)
}
\arguments{
\item{object}{Seurat object containing a polyAsiteAssay}

\item{assay}{Name of polyAsiteAssay}

\item{polyAdb.file}{Location to find polyAdbv3 file}

\item{max.dist}{Keep sites within this distance to a polyAdbv3 site. Default is 50 nucleotides.}
}
\value{
Seurat object with annotated meta.features for polyAsiteAssay. Keeps all features
in original matrix, but annotations columns will be populated with NAs if a feature does not
overlap with a polyAdbv3 polyA site within the distance specified my max.dist.
}
\description{
Overlap PAS with polyA db
}
\concept{annotation}
