% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{getting_geo_count_mat_for_age_prediction}
\alias{getting_geo_count_mat_for_age_prediction}
\title{Title of the function}
\usage{
getting_geo_count_mat_for_age_prediction(gse_id)
}
\description{
Description of the function.
}
