% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{getting_pdata_with_age_scores}
\alias{getting_pdata_with_age_scores}
\title{Title of the function}
\usage{
getting_pdata_with_age_scores(
  ES,
  filter_genes = FALSE,
  rank_norm = FALSE,
  REG = TRUE,
  PASTA = TRUE,
  CT46 = TRUE
)
}
\description{
Description of the function.
}
\examples{
library(magrittr)
library(jsutil)
library(data.table)
ES = getting_GEO_ES_for_age_model('GSE103938') \%T>\% pdim # 8113 8
pdata = getting_pdata_with_age_scores(ES)
dcast(pdata, treated_with ~ vector, value.var = 'PASTA', fun.aggregate = mean)
dcast(pdata, treated_with ~ vector, value.var = 'REG', fun.aggregate = mean)
data(ES_GSE103938)
ES = ES_GSE103938
pdata = getting_pdata_with_age_scores(ES, filter_genes = TRUE, rank_norm = TRUE)
dcast(pdata, treated_with ~ vector, value.var = 'PASTA', fun.aggregate = mean)
}
