% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{making_pseudobulks_and_predict_age}
\alias{making_pseudobulks_and_predict_age}
\title{Title of the function}
\usage{
making_pseudobulks_and_predict_age(
  seu,
  chunk_size = 1000,
  verbose = TRUE,
  REG = TRUE,
  PASTA = TRUE,
  CT46 = FALSE
)
}
\description{
Description of the function.
}
