% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurats_and_pseudobulks.R
\name{making_pseudobulks_from_seurat}
\alias{making_pseudobulks_from_seurat}
\title{Create Pseudobulk Samples from Seurat Object}
\usage{
making_pseudobulks_from_seurat(seu, chunk_size = 1000, verbose = TRUE)
}
\arguments{
\item{seu}{Seurat object.}

\item{chunk_size}{Numeric. Number of cells per pseudobulk sample.}

\item{verbose}{Logical. If TRUE, prints summary tables.}
}
\value{
Seurat object with pseudobulk aggregated data.
}
\description{
Aggregates single-cell data into pseudobulk samples based on cell type and age.
}
