% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{pasta}
\alias{\%&&\%}
\alias{\%&\%}
\alias{\%-\%}
\alias{\%.\%}
\alias{\%//\%}
\alias{\%_\%}
\alias{\%|\%}
\alias{pasta}
\title{Make pasta with strings}
\usage{
a \%_\% b

a \%-\% b

a \%.\% b

a \%//\% b

a \%&\% b

a \%&&\% b

a \%|\% b
}
\arguments{
\item{a, b}{characters.}
}
\description{
Pasta provides a set of easy to read infix function for the intuitive
pasting of characters.\cr
Currently the following functions are available for concatenation:

\%_\% pastes two strings by "_".

\%-\% pastes two strings by "-".

\%.\% pastes two strings by ".".

\%//\% pastes two strings by "/".

\%&\% pastes two strings by "".

\%&&\% pastes two strings by " ".

\%|\% pastes two strings with a "|".
}
\examples{
'macaroni' \%_\% 'sauce' # 'macaroni_sauce'
'ravioli' \%-\% 'recipe' # 'ravioli-recipe'
'spaghetti' \%.\% 'code' # 'spaghetti.code'
'fussili' \%//\% 'farfalle' # 'fussili/farfalle'
'pasta' \%&\% 'sciutta' # 'pastasciutta'
'noodle' \%&&\% 'soup' # 'noodle soup'
'carbonara' \%|\% 'ragù' # 'carbonara|ragù'
'spaghetti' \%&&\% 'al' \%-\% 'dente' # 'spaghetti al-dente'
}

