% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasta.R
\name{pASTA}
\alias{pASTA}
\title{pASTA for multi-phenotype analysis}
\usage{
pASTA(p.values, study.sizes, cor)
}
\arguments{
\item{p.values}{the p.value for each study.}

\item{study.sizes}{the sample size of each study.}

\item{cor}{the correlation matrix of the studies. For example, if each study
is independent, cor should be the idenity matrix}
}
\value{
a list containing the joint p value and the test statistic, which
  contains the optimal subset
}
\description{
description. TODO(youfei)
}
\examples{
# grab synthetic study for example
data("studies")
n.studies <- 5
study.sizes <- c(nrow(studies[[1]]), nrow(studies[[2]]), nrow(studies[[3]]),
                   nrow(studies[[4]]), nrow(studies[[5]]))
study.pvals <- rep(0, n.studies)
# Correlations of p-values among the studies.
# In this case the studies were generated independently so its just I
cor.matrix <- diag(1, n.studies)
# load the lrtest() function to conduct the likelihood ratio test
# Used just to generate the input p-values, not required in pASTA itself.

library(lmtest)

for(i in 1:n.studies) {
 # model with gene(G) by environment(E) interaction
 model <- glm(D ~ G + E + GbyE, data = studies[[i]], family = binomial)
 # model without G and GE interaction
 null.model <- glm(D ~ E, data = studies[[i]], family = binomial)
 # likelihood ratio test from the package lmtest
 study.pvals[i] = lmtest::lrtest(null.model, model)[2, 5]
}

pasta <- pASTA(study.pvals, study.sizes, cor.matrix)

pasta$p.pasta.joint
pasta$test.statistic$selected.subset
}
\references{
TODO(youfei)
}
