% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{PolyACoveragePlot}
\alias{PolyACoveragePlot}
\title{Coverage Plot for genes based on residuals}
\usage{
PolyACoveragePlot(
  object,
  assay = "polyA",
  gene = NULL,
  gene.names = "symbol",
  extend.downstream = 500,
  extend.upstream = 500,
  annotation = "transcript",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Name of polyAsite.Assay}

\item{gene}{Name of gene to plot. Will extract all}

\item{gene.names}{Column containing the gene where each polyA site is annotated.}

\item{extend.downstream}{}

\item{extend.upstream}{}
}
\description{
Calculate avergae length of tandem 3'UTRs
}
\concept{visualization}
