% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_Seurat_objects.R
\name{predicting_age_from_pseudobulks}
\alias{predicting_age_from_pseudobulks}
\title{Predict Age from Pseudobulk Data}
\usage{
predicting_age_from_pseudobulks(
  seu_bulk,
  REG = TRUE,
  PASTA = TRUE,
  CT46 = TRUE
)
}
\arguments{
\item{seu_bulk}{Seurat object with pseudobulk data.}

\item{REG}{Logical. Include REG model predictions.}

\item{PASTA}{Logical. Include PASTA model predictions.}

\item{CT46}{Logical. Include CT46 model predictions.}
}
\value{
data.table. Phenotype data with age predictions.
}
\description{
Uses processed pseudobulk data to predict age scores.
}
