% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{predicting_age_multiple_chunks}
\alias{predicting_age_multiple_chunks}
\title{Title of the function}
\usage{
predicting_age_multiple_chunks(
  seu,
  v_chunk_sizes = c(500, 1000),
  REG = TRUE,
  PASTA = TRUE,
  CT46 = TRUE,
  verbose = TRUE
)
}
\description{
Description of the function.
}
