% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{predicting_age_score}
\alias{predicting_age_score}
\title{Title of the function}
\usage{
predicting_age_score(mat, model_type = "PASTA")
}
\description{
Description of the function.
}
\examples{
library(magrittr)
data(ES_GSE103938)
ES = ES_GSE103938
mat = Biobase::exprs(ES)
mat \%<>\% filtering_age_model_genes_and_rank_norm
( v_age_scores = predicting_age_score(t(mat)) )
}
