% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{processing_geo_count_mat_for_age_prediction}
\alias{processing_geo_count_mat_for_age_prediction}
\title{Title of the function}
\usage{
processing_geo_count_mat_for_age_prediction(mat)
}
\description{
Description of the function.
}
