% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ReadPolyApipe}
\alias{ReadPolyApipe}
\title{The function reads in output from polyApipe and returns a count matrix}
\usage{
ReadPolyApipe(
  counts_file,
  peaks_file = NULL,
  sep = c(":", ",", ":"),
  verbose = TRUE
)
}
\arguments{
\item{counts_file}{file containing the counts from polyApipe in .tsv format}

\item{peaks_file}{gff file that requires the seqid, start, end, and strand of each peak}

\item{sep}{Separators}
}
\value{
Returns count matrix with peaks in the form "chromosome:start,end:strand"
such as "10:100000560,100000859,100000560,100000859:+"
}
\description{
The function reads in output from polyApipe and returns a count matrix
}
\concept{preprocessing}
