% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.R
\name{renaming_rows_in_mat_after_gene_id_conversion}
\alias{renaming_rows_in_mat_after_gene_id_conversion}
\title{Title of the function}
\usage{
renaming_rows_in_mat_after_gene_id_conversion(mat, dt_ids)
}
\description{
Description of the function.
}
